backend/ai.py
import os
from typing import List, Dict
from google.cloud import aiplatform
import vertexai
from vertexai.preview.generative_models import GenerativeModel, Tool

PROJECT_ID = os.getenv("PROJECT_ID", "mock-project-id")
LOCATION = os.getenv("LOCATION", "us-central1")

# Initialize Vertex AI
# vertexai.init(project=PROJECT_ID, location=LOCATION)

class AIController:
    def __init__(self):
        # In a real scenario, we load the Gemini model
        # self.model = GenerativeModel("gemini-pro")
        pass

    def generate_sql_query(self, user_query: str) -> str:
        """
        Uses Gemini to convert a natural language query into a SQL query for the Oracle DB.
        """
        prompt = f"""
        You are a SQL expert. Convert the following user query into a SQL query for a table named 'INVENTORY'.
        The table schema is:
        - ID (NUMBER)
        - NAME (VARCHAR)
        - CATEGORY (VARCHAR) -- e.g., 'Boys', 'Girls', 'Men', 'Women'
        - SIZE_VAL (VARCHAR) -- e.g., '7', '10', 'M', 'L'
        - COLOR (VARCHAR)
        - PRICE (NUMBER)
        - IMAGE_URL (VARCHAR)

        User Query: "{user_query}"

        Return only the SQL query.
        """

        # response = self.model.generate_content(prompt)
        # return response.text.strip().replace("```sql", "").replace("```", "")

        # Mock response for the demo
        print(f"AI generating SQL for: {user_query}")

        # Simple rule-based mock for the specific demo query
        if "shoes" in user_query.lower() and "blue" in user_query.lower():
            return "SELECT * FROM INVENTORY WHERE CATEGORY = 'Boys' AND SIZE_VAL = 7 AND COLOR = 'Blue'"

        return "SELECT * FROM INVENTORY WHERE ROWNUM <= 5"

    def interpret_search(self, user_query: str):
        """
        Alternative: Uses Gemini Function Calling or simply extracts parameters
        to pass to a structured search.
        """
        # Logic to extract keywords using Gemini
        return {"category": "Boys", "item": "Shoes", "size": 7, "color": "Blue"}
